backend/requirements.txt
# Django and Core Dependencies
Django==5.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0

# Database Support
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional)

# Firebase/Firestore Integration
firebase-admin==6.5.0
google-cloud-firestore==2.16.0

# Validation and Utilities
python-decouple==3.8  # For environment variables
Pillow==10.4.0  # Image processing
python-dateutil==2.8.2

# Development and Testing
pytest==8.3.2
pytest-django==4.8.0
coverage==7.6.1
factory-boy==3.3.0  # Test data generation

# API Documentation
drf-spectacular==0.27.2  # OpenAPI 3 schema generation

# Production Dependencies
gunicorn==22.0.0  # WSGI server
whitenoise==6.7.0  # Static file serving
sentry-sdk[django]==2.12.0  # Error tracking

# Development Tools (optional)
django-debug-toolbar==4.4.6
django-extensions==3.2.3
